use a shared runner source code for AOTI and ET
